<연차관리>[Fix]전자결재 도메인에서 필요한 부분 연동

+++ IAmGroupee/src/main/java/com/kh/iag/user/entity/UserDto.java
@@ -27,8 +27,8 @@ public class UserDto {
 	private String profile;	
 	private int alvCount;	
 	private int mlvCount;	
-	private float alvUsedCount;	
-	private float alvLeftCount;	
+	private double alvUsedCount;	
+	private double alvLeftCount;	
 	private int alvAddCount;	
 	private int alvTotalCount;	
 	private String personnelRight;
